Rather bothersome test validating urllib.
This is why we typically want to isolate network calls and prints are a bit of a pain to test. The work around popping off the list allows us to return different values for each call to , which stubs out the initial index call then the subsequent detail view. We then capture stdout to validate the print command returns what we think it should.